feat: initial release
This commit marks the initial release of the theme-class-provider library. theme-class-provider is a
React library designed to simplify theme management in React applications. It provides a
ThemeClassProvider component based on a list of classes provided by the user, allowing seamless
toggling throught the classlist in a cyclic manner. Additionally, it offers a hook to access the
theme context within components, making it easy to incorporate theme-related functionality into
React applications. With this initial release, users can now leverage theme-class-provider to
enhance the theming experience in their React projects. The library aims to streamline the process
of implementing and managing themes, providing a straightforward solution for developers seeking to
customize the appearance of their applications. This release sets the foundation for future updates
and improvements to the library, and we're excited to see how it will be adopted and integrated into
various projects within the React community. Contributions, feedback, and suggestions are welcome as
we continue to evolve and refine theme-class-provider to better serve the needs of React developers.